Speedup MemOutputArchive::save(std::string)
When reverse is enabled, we regularly take an in-memory snapshot of the MSX
machine. On small systems, the cost of enabling reverse is still significant.
When I profile the performance of the in-memory snapshots, then the compression
is by far the biggest cost (I'm not sure how to improve upon that). Another
significant subtask is serializing string objects. This patch make a small
speeds up in that step.

The in-memory format of a serialized string is:
 - the length of the string
 - followed by that many bytes for the actual string data
(In principle strings can have embedded zeros, so we can't use 0-termination).
Before this patch the code to (de)serialize a string was very simple: 'store
length' followed by 'reserve space and copy data'.

We don't know upfront how much memory an in-memory snapshot will require, so
it's possible we initially reserve a too small buffer and then during
serialization we need to reallocate a bigger one. Revision 13161 makes sure
that these reallocations are very rare. But nevertheless for correctness we
still have check for buffer overflow each time we store something new.

Before this patch, storing the length and reserving space for the string would
check twice for buffer overflow. Now we only reserve space once, and that space
will hold both the length and the actual string data.

This patch also inlines the OutputBuffer::allocate() and
OutputBuffer::deallocate() methods because they are both very small, they are
only called from one or two locations (one location is this string
serialization) and it does make a measurable performance difference.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13179 7512d703-071c-0410-b5c2-d2bdd895b7d1