Made MemBuffer objects moveable
A MemBuffer is a thin wrapper around an allocated chunk of memory. For example
it makes sure that the memory is deallocated again once the object goes out of
scope. MemBuffer objects cannot be copied (because that would lead to
double-free bugs). So when we want to return a MemBuffer from some function
we (used to) return a heap-allocated object by unique_ptr.

The internal representation of a MemBuffer object is basically just a pointer
plus a size. So it's a bit wasteful to allocate such small objects on the heap
(the pointer to the heap object is almost as large as the object itself).

In c++11 we can make non-copyable but moveable objects. This allows to
move-around MemBuffer objects (thus e.g. return it from a function) without
risking ever having two MemBuffer objects managing the same memory chunk. And
with this we can avoid the small heap-allocation.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13180 7512d703-071c-0410-b5c2-d2bdd895b7d1