Made TTFFont objects moveable
This patch is similar to the previous one. A TTFFont object is a thin wrapper
around a 'TTF_Font*'. It makes sure this pointer points to a properly
initialized font and it cleans up the font once it's no longer needed. TTFFont
objects cannot be copied. But because the internal representation is only a
single pointer, it's a shame to have to heap-allocate TTFFont objects.

Making TTFFont objects moveable allows to have both safety and efficiency.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13181 7512d703-071c-0410-b5c2-d2bdd895b7d1