Improve openGL wrapper objects
The classes Texture, FrameBufferObject and PixelBuffer are wrappers around the
corresponding openGL concepts. Basically these objects just wrap an integer
that is a handle for the openGL concept. They manage the lifetime of this
concept (properly initialize and cleanup) and add some convenient helper
methods.

Because these classes internally store just one int, it's a shame to have to
allocate these objects on the heap. Similar to the previous patches, this patch
makes the objects movable, but not copyable. And that allows to pass them around
as values while still protecting against double uses of the same resource.

--

Apart from turning these wrapper into value objects, this patch also makes a
slight change in the Texture class. Initially this class could handle 1D, 2D
and 3D texture types (or any other special texture type). It stored the type of
the texture in the wrapper object. Though we only ever used 2D textures (I
think way in the past we also used 1D textures). So now I've removed this type
member and always assume 2D textures.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13183 7512d703-071c-0410-b5c2-d2bdd895b7d1