Fixed cassette auto-run for some MSX models.
Previous implementation appeared not to work on e.g. a Panasonic
FS-A1WSX. This was due to the timing, i.e. when the command was typed
into the running MSX. It appeared that there was no good choice for the
delay to make it actually work. Thanks to BiFi, I was able to implement
a solution based on a break point (just before the prompt is shown).
This means the timing will always be OK. In order to prevent problems
with CLOAD and RUN, the implementation now puts the loading instruction
directly into the keyboard buffer, instead of using the type command.
This is also slightly faster.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13186 7512d703-071c-0410-b5c2-d2bdd895b7d1