Fixed assert on 'after time 5 reset'
After-emutime scheduled commands are executed at an exact emutime. This could
also be e.g. in the middle of a Z80 instruction. Some commands, like 'reset'
couldn't handle this and they trigger an assert.

Actually if the assert in the reset command is removed, everything does work as
it should. I also can't immediately think of any other command that will go
wrong when executed in the middle of an instruction. Though instead of removing
the assert, I decided to postpone the command till the next instruction
boundary. This makes it much easier to reason about the correctness of the
code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13189 7512d703-071c-0410-b5c2-d2bdd895b7d1