Use unique_ptr iso shared_ptr in AfterCommand
In C++98 only copyable smart-pointers could be stored in STL containers. So no
std::auto_ptr's, but shared_ptr's were OK. C++11 added move-semantics and this
makes it possible to also store unique_ptr's in containers.

In the AfterCommand class we had a vector of shared_ptr's. This was only to
work around the above limitation (we didn't actually need the shared-ownership
stuff). This patch replaces it with cheaper unique_ptr's.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13190 7512d703-071c-0410-b5c2-d2bdd895b7d1