Increase the serialization-version of HardwareConfig
Revision 13177 changed the HardwareConfig::config member variable type from
'unique_ptr<XMLElement>' to 'XMLElement'. This has a (small) impact on the
generated XML serialization output: the 'id' attributes in some XML tags are
gone now.

To load old savestates this doesn't matter (the attributes are simply ignored),
so we didn't need bw-compat serialization (load) code. However when a new
savestate is loaded in an older version, then the 'id' attributes will be
searched for and the savestate fails to load. It's of course much nicer if the
loading fails with an error about an unsupported savestate version than with an
error about missing stuff in the XML file. So revision 13177 should have
increased the serialization version.

I've increased the version in this revision. Unfortunately savestates created
between this revision and 13177 will still not give a proper error message when
loaded in an older (but not too much older) openMSX version.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13197 7512d703-071c-0410-b5c2-d2bdd895b7d1