Increase CAS-based baudrate for generated WAV files.
When I increased that factor years ago (July 11, 2005, revision 4558) to
4.6, higher values didn't work. But now I tried again and up to 6.2
works fine... No idea why... And I can't really figure out what the
factor means, in the sense that I have no idea what the final baudrate
is. It's probably not 6.2*1200... that sounds a bit much. And half of
that sounds like a bit small. And the net loading time improvement is
now more like a factor of 4.7 (3.6 with the old value.), so I'm
confused. (And yes, probably this commit message is not the best way to
communicate this ;-)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13198 7512d703-071c-0410-b5c2-d2bdd895b7d1