Don't crash when reverse'ing to 'saving to tape'
OpenMSX would crash when you reverse to a moment in time where the MSX is
saving something to tape. This patch fixes the crash.

We still don't support savestates (and thus also reverse) where the MSX is
saving to tape. When you do load such a savestate we force the cassetteplayer
to 'STOP' mode. The MSX software will then continue saving, but nothing is
actually written to the tape image. We do print a warning when this happens.

Reverse in combination with saving to tape is hard to fully implement because
it's difficult to control (e.g. also reverse) the state of the tape images on
the host filesystem.

I don't think reverse + saving to tape is an essential combination, so fully
implementing this is low priority for me. Of course it is important that we
don't crash.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13199 7512d703-071c-0410-b5c2-d2bdd895b7d1