When setting the renderer fails, try smaller scale_factors before giving up.
Before this patch, it was possible to set a scale_factor that would not
give an actual screen in SDL, meaning that openMSX wouldn't start up
anymore. You had to edit the settings.xml file manually to get out of
that situation. After this patch, openMSX will reduce the value of the
setting itself.
See also http://www.msx.org/forum/msx-talk/openmsx/catapult-081-linux-mint
where this was first reported.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13200 7512d703-071c-0410-b5c2-d2bdd895b7d1