Made it possible to 'bind' or 'after' 'mouse motion'
This patch makes it possible to do:
  bind "mouse motion" { do some stuff }
  after "mouse motion" { do some stuff }
Before it was only possible to use 'after' on a specific mouse motion event
(like 'mouse motion 123 -456), but that's of course not very practical. Using
'bind' for such specific events was not (and still is not) possible.

One of the following patches will use this new feature to improve the OSD menu
(Android touch events are shown as mouse events by SDL).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13201 7512d703-071c-0410-b5c2-d2bdd895b7d1