Fixed hang on 'disasm 0x100 bla.txt'
Apparently in a Tcl script it's not an error to compare an integer with a
(non-integer) string. Instead the integer is converted to string and a string
comparison is performed. (AFAIU this is because Tcl is a weakly typed language
and both integers and strings can be compared using the same '<' operator
syntax).

A consequence of this is that when you execute the command
    disasm 0x100 bla.txt
openMSX will hang. The 2nd parameter is supposed to be an integer. The disasm
command is a Tcl proc that internally has a for loop like this:
    for {set i 0} {$i < $num} {incr i} {
And 'num' is the name of that 2nd parameter. So in this scenario, that
comparison always returns true and openMSX hangs in an infinite Tcl loop.

Fixed by changing the loop to
    for {set i 0} {$i < int($num)} {incr i} {
This forces a conversion of the 2nd parameter to an integer and when that
fails you get a nice error message like:
    expected number but got "bla.txt"

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13207 7512d703-071c-0410-b5c2-d2bdd895b7d1