Implemented much improved command engine timing
Timing is based on the information in the document
  doc/vdp-vram-timing/vdp-timing.html

This is still a bit experimental. I did some testing myself and it seems to
work fine (with excellent timing results for e.g. the LINE command). So it's
ready for wider testing (I mean by more people). But I can't guarantee there
won't be any regressions.

Compared to the old code, this version runs _much_ slower. Partly because I've
undone many of the optimizations that were implemented over the years and
partly because the time-bookkeeping is rather simplistic ATM. Hopefully both
points can be addressed in the future. But I already wanted to check in a
simple version of the code.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13225 7512d703-071c-0410-b5c2-d2bdd895b7d1