Remove use of videosource setting from Display
The goal of this and the next few patches is to make 'videosource' a per-machine
setting. This patch removes the use of this setting from all global objects (so
from all places where we can't easily get access to a per-machine setting).

Actually the only global user was the Display class. It used 'videosource' for
two different reasons:
1) Display listened to videosource changes to check whether the renderer needs
   to be recreated on a setting change. I don't see why this is (still) needed
   (and the comments in the code already hinted at this obsolete behavior). So
   I just removed this part.
2) The other use was to check whether a received finish-frame-event should
   actually trigger a display redraw (e.g. you don't need to redraw when a
   frame for a currently invisible layer is finished).

I solved 2) by putting more context info in the finish-frame-event itself. So
now when you receive such an event, from the event itself it's clear whether
for example this frame can be ignored for rendering or maybe is part of an
image that consists out of two superimposed frames.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13232 7512d703-071c-0410-b5c2-d2bdd895b7d1