Transfer per-machine settings on reverse
After making 'videosource' a per-machine setting, when you e.g. select the
gfx9000 videosource and you reverse, the msx video source would be re-selected.

The reason for this was that non-saved per-machine settings would revert to
their default value on reverse. (Non-saved settings are settings whose value
doesn't end up in the settings.xml file, so whose value is lost on an openmsx
restart).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13234 7512d703-071c-0410-b5c2-d2bdd895b7d1