Improve the CliComm setting messages after the recent videosource changes
These changes were made based on GuyveR800 feedback.

The videosource setting is the only 'dynamic' setting we currently have.
Changing it from a global to a machine-specific setting caused some fallout
in the CliComm setting announcement stuff. This patch tries to improve the
situation, but likely won't yet fix all problems.

There were two main issues:
1) The creation of the videosource setting was announced before the new msx
machine was announced. Since per-machine messages include the machine id. This
order could be 'confusing' for clicomm clients.
2) The videosource setting is the only setting we (currently) have that uses
the notifyPropertyChange() method (to announce the set of possible values has
changed). Though it was possible that first a new value (still impossible) was
announced before is was announced that this value is also an allowed value.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13236 7512d703-071c-0410-b5c2-d2bdd895b7d1