Correctly restore video9000 state on loadstate
The state in video9000 itself was correctly restored, but the state that
(indirectly) depends on this was not (see the Video9000::writeIO() and recalc()
methods).

This 'sort-of' fixes the reverse_frame display issue with video9000:
  #469 Screen not/wrongly updated while doing reverse_frame

Although the symptom is gone now, the code is very fragile. Video9000::recalc()
uses Display::findLayer() to lookup the correct layer by name. When there are
multiple MSX machines (this is temporary the case when doing reverse) then
there are multiple layers with the same name. And by luck the current code
always returns the correct layer (from the new machine) in case of reverse.

When using multiple MSX machine, each with a Video9000 extension, I guess it is
still possible to create a situation where this goes wrong. I didn't bother
changing this now, because it will (hopefully soon) be reworked to be able to
handle more dynamic videosources.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13237 7512d703-071c-0410-b5c2-d2bdd895b7d1