Remove findLayer() calls from Video9000::recalc()
This is a followup on revision 13237, see that commit message for why using
findLayer() is bad.

Next to the stuff mentioned in rev13237, using findLayer() also won't work in
case there are multiple V99x8 (or V9990) chips in one MSX machine (we'd like to
support this in the near(?) future).

This patch adds a device reference to the V99x8 in the video9000 config file
(it already had a reference to the V9990). So even if there are multiple
V99x8/V9990 chips in the same machine, the config now refers to a specific
chip.

So once the Video9000 device knows about the VDP/V9990 device, it can query
that device for the specific Layer (=PostProcessor) it is rendering to. And
thus this allows to get rid of the findLayer() calls with hardcoded layer
names.

Although this change is conceptually simple, this patch is quite large because:
- The new getPostProcessor() call has to pass through 2 indirection layers
  (renderer and rasterizer).
- Each indirection has an abstract base class, sometimes with a dummy
  implementation.
- This change is required on both the VDP and V9990 side.
- All these changes touch both the .hh and .cc file

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13238 7512d703-071c-0410-b5c2-d2bdd895b7d1