Report error on invalid per-machine setting value
Report an error when a per-machine setting is given an invalid value. For
example try "set PSG_volume bla". Now this gives an error that an integer value
is expected for this setting.

I actually discoverd this problem while working on the videosource stuff (this
has recently become a per-machine setting). But because this is an independent
change I've made it a separate patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13239 7512d703-071c-0410-b5c2-d2bdd895b7d1