Made videosource setting values dynamic
This patch removes the fixed set of enum values for the videosource setting and
replaces it with dynamically (un)registered names.

In each location in the source where the VideoSource enum was used, we now use
either an ID (an int) or a name (a std::string). This patch also refactors some
tests against hardcoded VideoSource values (typically VIDEO_LASERDISC or
VIDEO_9000).

This patch does break video recording. This will be restored in a follow-up
patch.

It is not yet possible to use multiple V99x8/V9990 chips. But we're getting
close. AFAICS now, the only remaining todo is to dynamically generate
(per-machine unique) videosource names.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13240 7512d703-071c-0410-b5c2-d2bdd895b7d1