Fix video recording
The previous patch (intentionally) broke video recording. This patch restores
that functionality.

This is a relatively small patch. Initially I though it would be much harder.
That's why I postponed this till a later patch. I could have merged this patch
into the previous one, but having it separate is not too bad either.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13241 7512d703-071c-0410-b5c2-d2bdd895b7d1