Don't crash when using multiple RTC chips
The RP5C01 chip creates the (per-machine) 'rtcmode' setting. When using multiple
such chips in one MSX machine this triggered an assertion because we can't have
multiple settings with the same name.

This patch renames the setting to "<device-name> mode", with this exception:
  "Real time clock" -> rtcmode
And thus similar to the VDP naming stuff, this exception is by far the common
case.

Of course it's not very useful to have more than one RTC chip in one MSX
machine, but I triggered this crash when experimenting with the 'Neos MA-20'
extension (a MSX1->MSX2 'upgrade' extension which includes a RTC).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13248 7512d703-071c-0410-b5c2-d2bdd895b7d1