Use tiger-tree-hash (tth) to checksum HD images
Before we used 'sha1' to calculate a checksum for harddisk images. This patch
changes that to 'tth'. This checksum is (only) used to verify on loadstate
whether the content of the image has changed compared to when the savestate was
created. Note that reverse internally uses savestates.

The motivation for this change is the following:
 - reverse enabled
 - (regular) MSX writes to the HD image
Reverse regularly creates in-memory savestates, so when the HD content has
changed, this requires recalculating a new checksum. For even a moderately
sized harddisk image a sha1 calculation of the entire HD takes a noticeable
amount of time. (I hardly ever use emulated HD images, and even more rarely
write to them. So I wasn't aware of this problem. Thanks BiFi, GuyveR800,
Quibus for pointing me to this issue. And then it still took some months
before I actually did something about it ;-)

So this patch replaces sha1 with tth. The tth calculation itself is of
comparable speed as the sha1 calculation (even slightly slower). However tth is
a hash-tree. This means that the hash can be recalculated incrementally, so
when only a small portion of the data changes, the final hash value can be
recalculated _a_lot_ faster.

--

Next to harddisk image checksums we also use sha1 hashes for the filepool
stuff. But harddisk images are not searched for in a filepool, so this patch
doesn't touch that code. Would it be a good idea to also use tth for roms and
disks? For roms I don't think there's an advantage, for disks there might be.
But disks are small enough so that a full sha1 calculation doesn't hurt that
much. If we decide to change this, we also have to think about migrating the
filepool stuff.