Renamed ChangeLog to ChangeLog.old
We don't update ChangeLog anymore (for quite some time already). Instead we now
write the info in the (svn/git) commit messages.

I wonder if we should instead delete the ChangeLog file. Though for older
revisions this file contains more details than the commit messages, so it might
still be useful. Of course it's possible to retrieve the content of a deleted
file, but it's not obvious to go look for info in an invisible file .. not sure
what the best option is.