[1/7] cleanups: nullptr tests
  I have a local git branch with c++11 experiments. I cannot push that branch
  upstream yet because not all c++ compilers already support all those new
  c++11 constructs. To make it easier for me to maintain my local branch, in
  this patch series I did extract all the changes that can already be pushed.
  Some of these changes are not c++11 related.

There are two main ways to test whether a pointer is a null pointer in c++.
  if (recorder == nullptr) throw Exception("Failed to create recorder");
or
  if (!recorder) throw Exception("Failed to create recorder");
Both mean exactly the same, so choosing between them it's a matter of personal
coding style. Most of our code already uses the second variant. This patch
changes the rest of the code to also use that variant.

As I said this is just a coding style. In this paragraph I'll briefly motivate
why I personally prefer the 2nd variant. The examples above read as:
 1) if recorder is a null pointer then error
 2) if not recorder then error
The intention of the code is "if there is no recorder object then signal an
error". To me the 2nd variant more directly expresses that intention.

I vaguely remember that vc++ issues warnings on code constructs like
  bool myFunction() { return myPtr; }
so I did keep the explicit nullptr test there
  bool myFunction() { return myPtr != nullptr; }