[2/7] cleanups: algorithm stuff
c++11 lambda functions make it more practical to use STL algorithms (in this
patch, most often the find_if algorithm). Using algorithms make the code easier
to understand because they replace the intention of the code that is originally
only implicit present in the control flow with an explicitly named algorithm
with well-known semantics.

  There are likely many more locations in the code that can benefit from using
  STL algorithms. I didn't actively search for such locations. This patch just
  changes the ones that were present in my local c++11 branch.