move (iso copy) out of STL containers
Replace this pattern:
	string s = v.front();
	v.pop_front();
with:
	string s = move(v.front());
	v.pop_front();

In the first sequence the string is copied and then the original is destroyed.
In the replacement sequence the string is instead moved. This means that 'the
internal guts' of the original string are stolen and given to the new string,
the original string is left in an 'unspecified but safe to delete' state.

This example used string/front()/pop_front() but other combinations are
possible as well. The openMSX code actually already used this move-out pattern
in a few locations.

None of these changes are in performance critical code. I don't expect any
measurable performance differences. I did see a (small) code size reduction.