change configs: <name>/hardwareconfig.xml -> <name>.xml
Before this patch, the machine/extension configuration files were organized
like this:
  share
    machines
      Philips_VG_8235
        hardwareconfig.xml
      Panasonic_FS-A1GT
        hardwareconfig.xml
      ...
    extensions
      ram1mb
        hardwareconfig.xml
      ...
So each machine or extension had it's own subdirectory with (mostly) one file
'hardwareconfig.xml' in that subdir. This patch changes the structure to:
  share
    machines
      Philips_VG_8235.xml
      Panasonic_FS-A1GT.xml
      ...
    extensions
      ram1mb.xml
      ...
So each directory is replaced with a file that is directly the machine or
extension configuration.

First some (historical) background info:
  This structure with the extra subdirectory was created about 11-12 years ago.
  At that time, next to the hardwareconfig.xml file, we also put the system
  roms for the machine/extension in that subdir. I'm not 100% sure but I think
  the intention was to also include meta-information in this subdirectory. So
  at the time this extra directory made sense.

  Later we introduced the filepool stuff (also already 8-9 years ago), and we
  put meta-info in the hardwareconfig.xml file itself. So now (with a few
  exceptions) the specific machine/extension subdirs only contained one file
  anymore.

Motivation for this change:
1) Even though the filepool stuff was introduced a long time ago, occasionally
we still hear from users who try to put their system roms in the machine/
extension specific subdirs. Maybe removing these subdirs will discourage this
unnecessary work. (If only users would read the documentation ;-)

2) Startup speed. During the startup sequence, openMSX scans the list of
available machines (e.g. to create the list of allowed values for the 'machine'
setting). When there is an extra subdirectory, this scanning takes a bit more
time. On a desktop machine this time is insignificant. But on embedded device
(e.g. Dingoo) starting openMSX takes multiple seconds, every small bit we can
remove from this time is useful.
It's actually not that easy to measure the speed gain from this patch. Regular
benchmarks show that the machine scan routine is (very approx) twice as fast
and that this routine takes about 1% of total startup time. BUT this is with
all the filesystem data cached by the OS. The more interesting case is when the
OS has to actually read this config stuff from storage (this will often be the
case for embedded systems). I suspect the speed gain will be bigger in this
case.

So this patch changes the config structure and adapts the code. But for
backwards compatibility the old structure is also still recognized.

I left the organization of the CBIOS machines unchanged (so still with the
extra subdir) because for those machine we do ship systemroms. An alternative
could have been to create a default filepool for CBIOS.