Eliminate two (potential) sources of input event latency
Thanks to dunnius/Quibus for providing the scenario that uncovered the problem
(to me):
 1) Go to tas mode (command: set mode tas).
 2) Keep the left cursor key pressed for some time and then pause emulation.
 3) Go back to some point in time where the key was already pressed (e.g. by
    clicking in the reverse bar). Note the current frame number.
 3) Truncate the reverse-event-log (command: reverse truncatereplay)
 4) Re-press the left cursor key and advance a few frames (I did this by
    binding the 'advance_frame' command to some key).
  --> The expectation is that the left cursor key remains pressed in the event
      log.
 5) Save the replay (command: reverse savereplay /tmp/bla).
  --> If you study the event log in the saved file you notice that the left
      cursor key was released for a short duration.
 6) Go back to the frame number of step 2 (e.g. with the command
    reverse_frame).
  --> Also the 'toggle_cursors' script (activated by default in tas mode) show
      that the cursor key is not pressed at this point.

I found two possible causes for this problem:

a) EventDelay

The EventDelay class intentionally delays input events for a (short) amount of
time. It was introduced over 9 years ago to solve the problem of 'lost input
events when typing too fast' (see doc/missed-key.dia for some more info). Today
I can only reproduce this input problem when I set minframeskip to a high value
and load my system pretty hard. So maybe this class has outlived its purpose?

Anyway, the problem was that even when input_delay was set to 0.0, there STILL
was some delay: the time between EventDelay received the events and the call
to EventDelay::sync() that actually distributes the events. There was just no
EXTRA delay.

So this patch bypasses the sync() call for the case that input_delay==0.0. I've
considered just dropping the EventDelay stuff. Maybe I'll still do that in a
later patch. (Though at some point in the future I'd like to work on netplay,
and something very similar to EventDelay is needed to hide network latency).

b) Event delivery

[ I'm not sure how much this part contributes to the original problem. It
  was the first thing I tried, but it didn't (fully?) fix this problem. ]

The EventDistributor accepts events from any thread and delivers them in the
main emulation thread (in a way it's our thread-save-message-passing
mechanism). When an event is submitted, the main thread needs 'eventually' be
interrupted from MSX emulation and do the delivery. It cannot be interrupted
immediately because that would introduce (expensive) thread-synchronization
stuff in the fast-path of the CPU emulation code. But this 'eventual' delivery
does of course contribute to latency.

In the special case that the event-generating-thread is the main thread we
don't require extra synchronization in the CPU emulation code and we can
immediately exit CPU emulation. So that's what this patch does. Luckily many
events actually are generated from the main emulation thread, including most
(but not all) input events.