[1/2] Remove the asm versions of the Scale2x routine
The next patch will re-add the routine in a modified form. This patch also
slightly changes the interface to the c++ version of this routine. That's
why I decided to split this change over two commits.

See commit message of the next patch for a more detailed motivation for this
change.

I hope I didn't make any mistakes in (manually) editing the visual studio
configuration files.