[2/2] Added SSE2 version of the scale2x algorithm using intrinsics
Before this patch series our scale2x asm routine was written in gnu inline
assembly or as separate .asm files for visual studio. So that are two versions
of the same routine. The latter even required separate variants for x86 and
x64. We also need separate versions for 16bpp and 32bpp. And then we still have
separate versions for MMX, SSE and SSE2. That are a lot of possible
combinations. But not all of these combinations are actually implemented.

The goal of this patch is to unify all those versions as much as possible into
a single routine. This is achieved via two mechanisms:

1) Drop support for MMX and SSE.

The first x86 CPUs supporting the SSE2 instruction set are the 'Intel Pentium4'
(2001) and 'AMD Athlon64' (2003). So by now the majority of all x86 CPUs that
are still in use should support SSE2. I don't think we loose much by dropping
MMX and SSE.

Over time when more (all?) of our asm routines are converted to SSE2, we can
also drop the HostCPU class that at-run-time tries to detect the supported
instruction set.

2) Use intrinsics instead of (inline) assembly.

SSE2 compiler intrinsics are special functions like _mm_or_si128() that (more
or less) directly map on x86 SSE2 instructions. So when you use these functions
you (more or less) tell the compiler which instructions to generate. However
compared to writing assembly, you do not have to specify in exactly what order
the instructions should be scheduled and exactly which registers they should
use. Also the control flow and address calculations can still be written as
normal c++ code.

Usually the compiler has a much better understanding of the processors execution
pipeline than the programmer (and this pipeline is also different for the
different subtypes of x86 CPUs) so usually the compiler is better at scheduling
instructions in an optimal order.

Intrinsics are much easier to use in inline functions and templates. In this
specific case this allowed to write a single core routine that can handle both
16bpp and 32bpp.

These compiler intrinsics have been around for a very long time already, so why
didn't we use them earlier? The problem is that this mechanism works a bit
different in visual studio and in gcc. In VS these intrinsics are always
available, whether the target CPU actually supports these instructions or not.
OTOH gcc only makes them available when compiling for a target CPU that does
support them. So VS allows to generate the code and requires a run-time check
to see if it's safe to call (that version of) the routine. When gcc generates
the code it's always safe to call the routine, but it's not possible generate
faster versions that are only optionally called at run time. (Later versions of
gcc do support this, but I didn't really look into this yet).

Gcc sets the #define __SSE2__ when the SSE2 intrinsics are available. Since we
dropped MMX/SSE, we no longer require a run-time check for this. So combined
this allows to simply put the intrinsics inside a '#ifdef __SSE2__' test
without extra run-time checks. For visual studio we need to manually set this

--

The trigger that got me interested again in this stuff is this blog post:
  http://markplusplus.wordpress.com/2007/03/14/fast-sse-select-operation/
Short summary: SIMD code often uses a 'select' operation to convert conditional
code into straight-line code. Very often (also in Intel papers) this select
operation is implemented as:
    (x0 & ~mask) | (x1 & mask)
But this is often a better way:
    ((x0 ^ x1) & mask) ^ x0
Both variants generate 3 x86 (SIMD) instructions, but the latter only requires
commutative operations and this has advantages on a 2-operand instruction set
(the destination register is the same as one of the two input registers).

So I remembered that our scale2x asm routine used the first select variant and
decided to take a look. And then I saw we didn't yet have an SSE2 version of
this routine (and I didn't want to tweak the MMX/SSE versions anymore).

Some _very_rough_ performance numbers (reported numbers are for full emulation
time, not only the scale2x routine itself):
  c++ version:      15.7s
  old SSE version:  15.1s
  new SSE2 version: 13.3s
This is a much bigger performance improvement than I expected. Possibly because
the legacy SIMD instructions are not well optimized anymore in modern x86
CPUs???

And of course modern x86 CPUs are more than fast enough to just run the C++
version of the scale2x routine. So we could just drop this asm routine
completely. Though I find playing with this stuff interesting and maybe in some
cases it can help to extend the battery life of your laptop.