Replace asm Scanline routine with intrinsics
This is similar to the recent revision 'baa1ca', see that patch for the
motivation.

For 32bpp we already had a SSE2 routine and the new intrinsic code is basically
using the same algorithm. When I measured old/new implementation, I saw a tiny
advantage for the new version (average of about 20 test runs), though it was
still within the measurement noise. (Or the difference could be because of
better instruction scheduling).

For 16bpp the old routine was MMX-ext only, now it's also SSE2. I expect this
new routine to be faster, but I couldn't find in a few minutes how to disable
VSYNC for the SDLGLFB-16 renderer on my system. So I only checked correctness.

Though the main purpose of this patch was not to improve performance, but to
simplify the code. This patch removes netto about 380 lines of code.