Show current value of TAS-widgets in the middle of a frame
Before this patch, the TAS-widgets (frame-count, ram-watch, cursor-keys) were
only updated once per frame (at the start of the frame). If the emulation is
paused this can be in the middle of a frame. This can lead to the following
situation:
 - start of frame, ram-watch peeks for the to-be-displayed values
 - a little later the MSX writes new values to memory
 - still a bit later, but before the start of the next frame, emulation
   is paused.
In this scenario the displayed ram-watch-values and the actual ram-values are
different.

This patch also updates the TAS-widgets at least 10 times per second. So it
e.g. triggers when emulation is paused, but also on very slow emulation (set
speed 1).

I've also extended the frame-count-indicator to show whether we're at the
start of a frame or not.