Fix part of bug #470
openMSX-debugger crashes when the OPL3Cartridge2 extensions are used
  https://sourceforge.net/p/openmsx/bugs/470/

See also the corresponding commit in the openmsx-debugger-repo.

In case the OPL3Cartridge2_{mono,stereo} extension is inserted, the
'guess_title' script generates an error. Internally the script does something
like this:
   set device [machine_info slot $ps $ss 1]
   set device_info [machine_info device $rom]
So the scripts queries which device is present is page 1 of a certain (sub)slot
and then queries more info on that device. This generally works fine, though in
case of the OPL3Cartridge2 extensions, the same 16kB page contains both a ROM
and a RAM part. So the first query does not return a single device name and
then the second query goes wrong.

This patch only adds a quick hack in the 'guess_title' script to not stop
when the 2nd query fails. A proper fix would be to make the 1st query return
a proper Tcl list and adjust all our scripts to deal with the lists. I'll do
that in a follow-up patch.