Make 'machine_info slot' return a list of devices
This is a follow-up patch on the #470 bug fix.

The machine info topic 'slot' returns the name of the MSX device that is
registered in the given slot/sub-slot/page combination (16kB pages). Though in
case there are multiple devices registered for (different parts of) this page,
we used to return a string that was the concatenation of the individual device
names. Unfortunately it wasn't formatted as a proper Tcl list, so it wasn't
possible to unambiguously extract the original names. This patch does change it
to a proper Tcl list.

When there is no device registered in a certain (sub)slot, we used to return
'empty' (historic artifact, so that it can directly be shown in the 'showmap'
command). No we return an empty list.

This patch also extends all our existing Tcl scripts that use this info topic
to properly handle lists. Some procs could already partly handle it, but procs
like 'guess_title' (before the previous patch) were unprepared.